# angularFE/bankingapp-frontend/src/app/model/clientproduct-request.ts
export interface ClientProductRequest {    
    clientId: string;
    productId: string;
    initialBalance: string;
    startDate: string;
    endDate: string;
    loanAmount: string;
    fixedInstallment: string;
}
  