# api-docs/clientproduct.api-docs.json
{"openapi":"3.0.1","info":{"title":"Client-product Service API","description":"This is the REST API for the Client-product Service","license":{"name":"Apache 2.0"},"version":"v1.0"},"servers":[{"url":"http://794db08b3007:8085","description":"Generated server url"}],"paths":{"/api/v1/client-products/loans/{clientProductId}":{"get":{"tags":["client-product-controller"],"operationId":"getLoanProductById","parameters":[{"name":"clientProductId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanProductResponse"}}}}}},"put":{"tags":["client-product-controller"],"operationId":"updateProductById","parameters":[{"name":"clientProductId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientProductRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/client-products/accounts/{clientProductId}":{"get":{"tags":["client-product-controller"],"operationId":"getAccountProductById","parameters":[{"name":"clientProductId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountProductResponse"}}}}}},"put":{"tags":["client-product-controller"],"operationId":"updateAccountProductById","parameters":[{"name":"clientProductId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientProductRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/client-products":{"get":{"tags":["client-product-controller"],"operationId":"getAllClientProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientProductResponse"}}}}}}},"post":{"tags":["client-product-controller"],"operationId":"createClientProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientProductRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/client-products/{clientProductId}":{"get":{"tags":["client-product-controller"],"operationId":"getClientProductById","parameters":[{"name":"clientProductId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientProductResponse"}}}}}}},"/api/v1/client-products/client/{clientId}":{"get":{"tags":["client-product-controller"],"operationId":"getClientProductsByClientId","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientProductResponse"}}}}}}}},"/api/v1/client-products/client/{clientId}/loans":{"get":{"tags":["client-product-controller"],"operationId":"getLoanProductsByClientId","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanProductResponse"}}}}}}}},"/api/v1/client-products/client/{clientId}/accounts":{"get":{"tags":["client-product-controller"],"operationId":"getAccountProductsByClientId","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountProductResponse"}}}}}}}}},"components":{"schemas":{"ClientProductRequest":{"required":["clientId","productId","startDate"],"type":"object","properties":{"clientId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"initialBalance":{"type":"number"},"loanAmount":{"type":"number"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"fixedInstallment":{"type":"number"}}},"ClientProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["ACCOUNT","LOAN","UNKNOWN"]},"lastChargeDate":{"type":"string","format":"date"}}},"LoanProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["ACCOUNT","LOAN","UNKNOWN"]},"lastChargeDate":{"type":"string","format":"date"},"originalAmount":{"type":"number"},"fixedInstallment":{"type":"number"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}}},"AccountProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["ACCOUNT","LOAN","UNKNOWN"]},"lastChargeDate":{"type":"string","format":"date"},"accountBalance":{"type":"number"},"startDate":{"type":"string","format":"date"}}}}}}